[Behat] Remove SimpleProductFormTrait
