+++ src/Sylius/Behat/Page/Admin/Product/SimpleProduct/CreateSimpleProductPage.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\Behat\Page\Admin\Product\SimpleProduct;
 
+use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Session;
 use Sylius\Behat\Page\Admin\Crud\CreatePage as BaseCreatePage;
 use Sylius\Behat\Page\Admin\Product\Common\ProductAssociationsTrait;
@@ -32,7 +33,6 @@ class CreateSimpleProductPage extends BaseCreatePage implements CreateSimpleProd
     use ProductMediaTrait;
     use ProductTaxonomyTrait;
     use ProductTranslationsTrait;
-    use SimpleProductFormTrait;
 
     public function __construct(
         Session $session,
@@ -81,18 +81,84 @@ class CreateSimpleProductPage extends BaseCreatePage implements CreateSimpleProd
         $this->getElement('side_navigation_tab', ['%name%' => $tabName])->click();
     }
 
+    public function specifyCode(string $code): void
+    {
+        $this->getElement('code')->setValue($code);
+    }
+
+    public function specifyField(string $field, string $value): void
+    {
+        $this->getElement($field)->setValue($value);
+    }
+
+
+    public function selectShippingCategory(string $shippingCategoryName): void
+    {
+        $this->changeTab('shipping');
+        $this->getElement('field_shipping_category')->selectOption($shippingCategoryName);
+    }
+
+    public function setShippingRequired(bool $isShippingRequired): void
+    {
+        $this->changeTab('details');
+
+        if ($isShippingRequired) {
+            $this->getElement('field_shipping_required')->check();
+
+            return;
+        }
+
+        $this->getElement('field_shipping_required')->uncheck();
+    }
+
+    public function isShippingRequired(): bool
+    {
+        return $this->getElement('field_shipping_required')->isChecked();
+    }
+
+    public function hasTab(string $name): bool
+    {
+        return $this->hasElement('side_navigation_tab', ['%name%' => $name]);
+    }
+
+    protected function getElement(string $name, array $parameters = []): NodeElement
+    {
+        if (!isset($parameters['%locale%'])) {
+            $parameters['%locale%'] = 'en_US';
+        }
+
+        return parent::getElement($name, $parameters);
+    }
+
     protected function getDefinedElements(): array
     {
         return array_merge(
             parent::getDefinedElements(),
-            $this->getDefinedFormElements(),
             $this->getDefinedProductMediaElements(),
             $this->getDefinedProductAssociationsElements(),
             $this->getDefinedProductAttributesElements(),
             $this->getDefinedProductTranslationsElements(),
             $this->getDefinedProductTaxonomyElements(),
             $this->getDefinedProductChannelPricingsElements(),
+            [
+                'code' => '[data-test-code]',
+                'enabled' => '[data-test-enabled]',
+                'field_shipping_category' => '[name="sylius_admin_product[variant][shippingCategory]"]',
+                'field_shipping_required' => '[name="sylius_admin_product[variant][shippingRequired]"]',
+                'form' => '[data-live-name-value="sylius_admin:product:form"]',
+                'product_translation_accordion' => '[data-test-product-translations-accordion="%localeCode%"]',
+                'side_navigation_tab' => '[data-test-side-navigation-tab="%name%"]',
+            ],
         );
+    }
+
+    private function waitForFormUpdate(): void
+    {
+        $form = $this->getElement('form');
+        sleep(1); // we need to sleep, as sometimes the check below is executed faster than the form sets the busy attribute
+        $form->waitFor(1500, function () use ($form) {
+            return !$form->hasAttribute('busy');
+        });
     }
 
     private function selectElementFromAttributesDropdown(string $id): void
