+++ src/Sylius/Behat/Page/Admin/Product/SimpleProduct/UpdateSimpleProductPage.php
@@ -36,7 +36,6 @@ class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProd
     use ProductMediaTrait;
     use ProductTaxonomyTrait;
     use ProductTranslationsTrait;
-    use SimpleProductFormTrait;
 
     public function __construct(
         Session $session,
@@ -129,6 +128,64 @@ class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProd
         $this->getElement('enabled')->check();
     }
 
+    public function specifyCode(string $code): void
+    {
+        $this->getElement('code')->setValue($code);
+    }
+
+    public function specifyField(string $field, string $value): void
+    {
+        $this->getElement($field)->setValue($value);
+    }
+
+
+    public function selectShippingCategory(string $shippingCategoryName): void
+    {
+        $this->changeTab('shipping');
+        $this->getElement('field_shipping_category')->selectOption($shippingCategoryName);
+    }
+
+    public function setShippingRequired(bool $isShippingRequired): void
+    {
+        $this->changeTab('details');
+
+        if ($isShippingRequired) {
+            $this->getElement('field_shipping_required')->check();
+
+            return;
+        }
+
+        $this->getElement('field_shipping_required')->uncheck();
+    }
+
+    public function isShippingRequired(): bool
+    {
+        return $this->getElement('field_shipping_required')->isChecked();
+    }
+
+    public function hasTab(string $name): bool
+    {
+        return $this->hasElement('side_navigation_tab', ['%name%' => $name]);
+    }
+
+    protected function getElement(string $name, array $parameters = []): NodeElement
+    {
+        if (!isset($parameters['%locale%'])) {
+            $parameters['%locale%'] = 'en_US';
+        }
+
+        return parent::getElement($name, $parameters);
+    }
+
+    private function waitForFormUpdate(): void
+    {
+        $form = $this->getElement('form');
+        sleep(1); // we need to sleep, as sometimes the check below is executed faster than the form sets the busy attribute
+        $form->waitFor(1500, function () use ($form) {
+            return !$form->hasAttribute('busy');
+        });
+    }
+
     protected function getCodeElement(): NodeElement
     {
         return $this->getElement('code');
@@ -138,7 +195,6 @@ class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProd
     {
         return array_merge(
             parent::getDefinedElements(),
-            $this->getDefinedFormElements(),
             $this->getDefinedProductMediaElements(),
             $this->getDefinedProductAssociationsElements(),
             $this->getDefinedProductAttributesElements(),
@@ -146,7 +202,14 @@ class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProd
             $this->getDefinedProductTaxonomyElements(),
             $this->getDefinedProductChannelPricingsElements(),
             [
+                'code' => '[data-test-code]',
+                'enabled' => '[data-test-enabled]',
+                'field_shipping_category' => '[name="sylius_admin_product[variant][shippingCategory]"]',
+                'field_shipping_required' => '[name="sylius_admin_product[variant][shippingRequired]"]',
+                'form' => '[data-live-name-value="sylius_admin:product:form"]',
+                'product_translation_accordion' => '[data-test-product-translations-accordion="%localeCode%"]',
                 'show_product_button' => '[data-test-view-in-store]',
+                'side_navigation_tab' => '[data-test-side-navigation-tab="%name%"]',
             ],
         );
     }
